requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
stripe==7.6.0
dj-rest-auth==5.0.1
django-allauth==0.58.2
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0
